relocate tech-mentors.js
